+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+**Internal**:
+
+- Add stub metric summaries. ([#2757](https://github.com/getsentry/relay/pull/2757))
+
 ## 23.11.1
 
 **Features**:
@@ -18,7 +24,6 @@
 - Support comparison operators (`>`, `>=`, `<`, `<=`) for strings in dynamic sampling and metric extraction rules. Previously, these comparisons were only possible on numbers. ([#2730](https://github.com/getsentry/relay/pull/2730))
 - Postpone processing till the global config is available. ([#2697](https://github.com/getsentry/relay/pull/2697))
 - Skip running `NormalizeProcessor` on renormalization. ([#2744](https://github.com/getsentry/relay/pull/2744))
-- Add stub metric summaries. ([#2757](https://github.com/getsentry/relay/pull/2757))
 
 ## 23.11.0